Add files via upload
I sorted out the incorrect amount of cards in the deck.
Adapted the code to handle Ace's and give the user the decision whether to take the Ace as a 1 or an 11. 
Also did this for the dealer, she will take the Ace as an 11 if it gives her a total of at least 16. If not, she will take it as a 1.