Add files via upload
I created a method that stores relevant information regarding the outcome of each hand in a database on my local computer. 
This database (called blackjack) is in a MySQL server, so I have used the MySQL python connector to be able to write SQL in this python program. 
The table in this database where the data is stored is called 'games'.